requirements.txt
# Background Removal Service - Dependencies
# Install in dedicated venv: /home/amello/bg-removal-env

# API framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.9

# Background removal (IS-Net)
rembg[gpu]>=2.0.57
onnxruntime-gpu>=1.19.0

# Segment Anything Model (SAM ViT-L)
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
torch>=2.4.0
torchvision>=0.19.0

# Image processing
Pillow>=10.0.0
numpy>=1.26.0
